+++ Front-with-React/VigyBag/src/pages/Dashboard/Dashboard.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Navbar from "../components/Navbar/Navbar";
+import Navbar from "../../components/Navbar/Navbar";
 
 function Dashboard() {
     return (
