+++ Front-with-React/VigyBag/src/pages/Home/Home.jsx
@@ -1,0 +1,20 @@
+import React from 'react';
+
+import background from "../../assets/images/background.png";
+import Navbar from "../../components/Navbar/Navbar";
+
+function Home() {
+    return (
+        <>
+            <div style={{ backgroundImage: `url(${background})`, backgroundSize: 'cover', height: '100vh', position: 'relative' }}>
+
+                <h1>Welcome to Vigy Bag</h1>
+                <h1>Your Eco-Friendly</h1>
+                <h1>Shopping Heaven</h1>
+
+            </div>
+        </>
+    );
+}
+
+export default Home;
